+++ gptme/init.py
@@ -11,7 +11,7 @@ from .models import (
     get_recommended_model,
     set_default_model,
 )
-from .tabcomplete import load_readline_history, register_tabcomplete
+from .readline import load_readline_history, register_tabcomplete
 from .tools import init_tools
 from .util import console
